myth-buster-ai/requirements.txt
#API
fastapi
uvicorn

#Env variables
python-dotenv

#AI
langchain
langchain-openai
langchain-community
langchain-tavily
langgraph

# Grpah png vizualisation
graphviz